Put Role-based code into CustomUserDetailsService
