+++ src/main/java/com/sysc4806app/configuration/CustomUserDetailsService.java
@@ -1,10 +1,13 @@
 package com.sysc4806app.configuration;
 
+import com.sysc4806app.model.Role;
 import com.sysc4806app.model.User;
+import com.sysc4806app.repos.RoleRepo;
 import com.sysc4806app.repos.UserRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
@@ -12,19 +15,29 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
 
+import javax.transaction.Transactional;
 import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Collections;
 import java.util.List;
 
-@Service
+@Service("userDetailsService")
+@Transactional
 public class CustomUserDetailsService implements UserDetailsService {
 
+    private final UserRepo userRepo;
+    private final RoleRepo roleRepo;
+
     @Autowired
-    private UserRepo userRepository;
+    public CustomUserDetailsService(UserRepo userRepo, RoleRepo roleRepo) {
+        this.userRepo = userRepo;
+        this.roleRepo = roleRepo;
+    }
 
     @Override
     public UserDetails loadUserByUsername(String name) throws UsernameNotFoundException {
 
-        User user = userRepository.findByName(name);
+        User user = userRepo.findByName(name);
         if (user != null) {
             return buildUserForAuthentication(user);
         } else {
@@ -32,8 +45,22 @@ public class CustomUserDetailsService implements UserDetailsService {
         }
     }
 
+    private Collection<? extends GrantedAuthority> getAuthorities(Collection<Role> roles) {
+        return getGrantedAuthorities(roles);
+    }
+
+    private List<GrantedAuthority> getGrantedAuthorities(Collection<Role> roles) {
+        List<GrantedAuthority> authorities = new ArrayList<>();
+        for (Role role : roles) {
+            authorities.add(new SimpleGrantedAuthority(role.getName()));
+        }
+        return authorities;
+    }
+
     private UserDetails buildUserForAuthentication(User user) {
-        return new org.springframework.security.core.userdetails.User(user.getName(), user.getPassword(), new ArrayList<>());
+        return new org.springframework.security.core.userdetails.User(
+                user.getName(), user.getPassword(), true, true, true,
+                true, getAuthorities(user.getRoles()));
     }
 
 }
