--- src/main/java/com/sysc4806app/services/RoleBasedUserDetailsService.java
@@ -1,56 +1,0 @@
-package com.sysc4806app.services;
-
-import com.sysc4806app.model.Role;
-import com.sysc4806app.model.User;
-import com.sysc4806app.repos.RoleRepo;
-import com.sysc4806app.repos.UserRepo;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.GrantedAuthority;
-import org.springframework.security.core.authority.SimpleGrantedAuthority;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-import javax.transaction.Transactional;
-import java.util.*;
-
-@Service("userDetailsService")
-@Transactional
-public class RoleBasedUserDetailsService implements UserDetailsService {
-
-    private final UserRepo userRepo;
-    private final RoleRepo roleRepo;
-
-    @Autowired
-    public RoleBasedUserDetailsService(UserRepo userRepo, RoleRepo roleRepo) {
-        this.userRepo = userRepo;
-        this.roleRepo = roleRepo;
-    }
-
-    @Override
-    public UserDetails loadUserByUsername(String name) throws UsernameNotFoundException {
-        User user = userRepo.findByName(name);
-        if (user == null) {
-            return new org.springframework.security.core.userdetails.User(
-                    " ", " ", true, true, true, true,
-                    getAuthorities(Collections.singletonList(roleRepo.findByName("ROLE_USER"))));
-        }
-
-        return new org.springframework.security.core.userdetails.User(
-                user.getName(), user.getPassword(), true, true, true,
-                true, getAuthorities(user.getRoles()));
-    }
-
-    private Collection<? extends GrantedAuthority> getAuthorities(Collection<Role> roles) {
-        return getGrantedAuthorities(roles);
-    }
-
-    private List<GrantedAuthority> getGrantedAuthorities(Collection<Role> roles) {
-        List<GrantedAuthority> authorities = new ArrayList<>();
-        for (Role role : roles) {
-            authorities.add(new SimpleGrantedAuthority(role.getName()));
-        }
-        return authorities;
-    }
-}
